+++ R/coloring.R
@@ -31,8 +31,8 @@ new_edge_coloring <- function(
       # Applying alpha levels and getting mix
       col <- colorRamp2(
         x = c(
-          grDevices::adjustcolor(col = col_i, alpha.f = alpha_i),
-          grDevices::adjustcolor(col = col_j, alpha.f = alpha_j)
+          alphacolor(col = col_i, alpha.f = alpha_i),
+          alphacolor(col = col_j, alpha.f = alpha_j)
         )
       )(seq(0, 1, length.out = n))
 
@@ -47,13 +47,13 @@ new_edge_coloring <- function(
       alpha_j <- vertex_alpha_j[j]
 
       # Edge level params
-      col_ij   <- grDevices::adjustcolor(col = edge_color[e], alpha.f = edge_alpha[e])
+      col_ij   <- alphacolor(col = edge_color[e], alpha.f = edge_alpha[e])
 
       # Applying alpha levels and getting mix
       col <- colorRamp2(x=
         c(
-          grDevices::adjustcolor(col = col_ij, alpha.f = alpha_i),
-          grDevices::adjustcolor(col = col_ij, alpha.f = alpha_j)
+          alphacolor(col = col_ij, alpha.f = alpha_i),
+          alphacolor(col = col_ij, alpha.f = alpha_j)
         )
       )(seq(0, 1, length.out = n))
 
@@ -179,6 +179,10 @@ color_formula <- function(x, col, alpha, env, type, mix = 1, postfix = NULL) {
   else if (length(col) != n)
     stop("`col` has the wrong length (", length(col), "). When passing a ",
          "vector it should be of length ", n, ".", call. = FALSE)
+
+  # If all are NA, then it is replaced
+  if (all(is.na(col)))
+    col[] <- grDevices::hcl.colors(1)
 
   # Checking alpha
   alpha <- if (missing(alpha))